Update office_floors field in CustomUser model
Subject: Allow `null` in CustomUser office_floors field

Body: A change has been made to the `office_floors` field of the `CustomUser` model. The field is now able to accept `null` values. This change was necessary to enable cases where a user may not be associated with any specific office floor.

As part of the change, a new migration file has been generated to effectively change the schema of the existing database, thereby reflecting the null acceptance of the field.

In addition, tests were added for several models to ensure the accuracy and reliability of the code. This not only supports our new changes but promotes robustness in our code base.

These adjustments increase the flexibility and applicability of our user model, facilitating a more accurate representation of our user base.